Strands-Agent/simpleAgent/requirements.txt
fastapi==0.128.0
bedrock-agentcore==1.2.0
uvicorn==0.40.0
pydantic==2.12.5
strands-agents==1.12.0
typing-extensions==4.15.0
strands-agents[ollama]